use cuda 12.5 for clang-generated ir
we initially used cuda 12.3 because clang compiler v20 was complaining that 12.5 is only partially supported, and we thought its causing a compilation issue regarding fatbinary. turns out it also happens with cuda 12.3 and we were able to fix it with additiong of --cuda-device-only flag, so the host code was not generated and hence no compilation issue regarding fatbinary.